Implement ListLiteral parser (#4)
Implemented a parser for the ListLiteral expression. Rule considers list of ',' separated expression values.
Need to revisit parsing logic for URL and File as it clashes with listLiteral parsing logic.
All visible character CharPredicate will not work in case of URL and File.
We need to write more specific Predicate.